Add support file for RC Controller
